refactor filecheck testing
