+++ python/triton/testing/__init__.py
@@ -1,0 +1,7 @@
+from .filecheck import run_filecheck_test, filecheck_test, run_parser
+
+__all__ = [
+    "run_filecheck_test",
+    "filecheck_test",
+    "run_parser",
+]
